settings: merge About into System
Combine the About and System settings sections into one:
- Check for updates moves to the top of System
- System's own rows (language, artifacts, terminal folder, cached
  media, telemetry consent, Doctor, developer tools) keep their place
- About's remaining rows (app identity fields, Account) move to the
  bottom of System under an "About" subhead

About is no longer a routable settings section. Both the "about"
and "updates" legacy routes now redirect to "system".
