+++ src/features/settings/ui/SystemSettings.tsx
@@ -1,7 +1,16 @@
 import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
+import { useQueryClient } from "@tanstack/react-query";
 import { toast } from "sonner";
 import { FolderOpen, RotateCcw, Terminal, Trash2 } from "lucide-react";
+import {
+  listAuthWorkspaces,
+  logout,
+  switchAuthWorkspace,
+  type AuthStatus,
+  type AuthWorkspaceList,
+} from "@/features/auth/api/auth";
+import { resetChatRuntimeStartup } from "@/app/lib/chatRuntimeStartup";
 import { cn } from "@/shared/lib/cn";
 import { getPlatform } from "@/shared/lib/platform";
 import { SettingsPage } from "@/shared/ui/SettingsPage";
@@ -24,6 +33,7 @@ import { type LocalePreference, useLocale } from "@/shared/i18n";
 import { useArtifactRootPreference } from "@/shared/artifacts/useArtifactRootPreference";
 import { useTerminalFallbackCwdPreference } from "@/features/terminal/lib/terminalCwdPreference";
 import { useProfileCapability } from "@/shared/profile/capabilities";
+import { UpdatesSettings } from "@/features/updates/ui/UpdatesSettings";
 import { RuntimeConfigSettings } from "./RuntimeConfigSettings";
 import { TelemetryConsentRow } from "./TelemetryConsentRow";
 import { DoctorSettings } from "./DoctorSettings";
@@ -33,6 +43,30 @@ import {
   installBbCli,
   type BbCliStatus,
 } from "@/shared/api/bbCli";
+
+interface AboutAppInfo {
+  version: string;
+  tauriVersion: string;
+  identifier: string;
+}
+
+function AboutInfoRow({ label, value }: { label: string; value: string }) {
+  return (
+    <SettingsRow
+      label={<span className="text-muted-foreground">{label}</span>}
+      action={
+        <span className="block min-w-0 truncate text-right text-sm">
+          {value}
+        </span>
+      }
+    />
+  );
+}
+
+interface SystemSettingsProps {
+  authStatus?: AuthStatus;
+  onLoggedOut?: (status: AuthStatus) => void;
+}
 
 // System (rev 3): "settings about Berd as installed software on this
 // machine" -- split out of the old GeneralSettings.tsx. Language lives here
@@ -56,8 +90,23 @@ import {
 // status line under the row's description, in the same style as the
 // destructive text health-check surfaces elsewhere use -- no dot, no
 // indent, and it's just additional description text, not a badge.
-export function SystemSettings() {
+//
+// Rev 5 (Aug 19): "About" merged into System instead of staying its own nav
+// destination -- both pages were "settings about Berd itself," just split
+// across install-level vs. identity-level, which wasn't a distinction
+// worth a second sidebar row. The embedded Updates card (app version + the
+// "Check for updates" row) moved to the very top of this page, since
+// checking for updates is the row people actually look for. The rest of
+// About's content -- app identity fields and Account -- moved down to the
+// bottom under an "About" subhead, in the same order they rendered on the
+// old About page. `about` and the legacy `updates` route both redirect
+// here now (see settingsSections.ts).
+export function SystemSettings({
+  authStatus,
+  onLoggedOut,
+}: SystemSettingsProps) {
   const { t } = useTranslation("settings");
+  const queryClient = useQueryClient();
   const { preference, setLocalePreference, systemLocaleLabel } = useLocale();
   const [bbCliStatus, setBbCliStatus] = useState<BbCliStatus | null>(null);
   const [bbCliLoading, setBbCliLoading] = useState(false);
@@ -65,10 +114,19 @@ export function SystemSettings() {
   const [clearCacheDialogOpen, setClearCacheDialogOpen] = useState(false);
   const [clearingCache, setClearingCache] = useState(false);
   const [doctorDialogOpen, setDoctorDialogOpen] = useState(false);
+  const [appInfo, setAppInfo] = useState<AboutAppInfo | null>(null);
+  const [loggingOut, setLoggingOut] = useState(false);
+  const [workspaceList, setWorkspaceList] = useState<AuthWorkspaceList | null>(
+    null,
+  );
+  const [workspaceLoading, setWorkspaceLoading] = useState(false);
+  const [workspaceError, setWorkspaceError] = useState(false);
+  const [workspaceSwitching, setWorkspaceSwitching] = useState(false);
   const artifactRootPreference = useArtifactRootPreference();
   const terminalFallbackCwdPreference = useTerminalFallbackCwdPreference();
   const doctorEnabled = useProfileCapability("doctor");
   const agentToolsEnabled = useProfileCapability("agentTools");
+  const updatesEnabled = useProfileCapability("updates");
   const isMac = getPlatform() === "mac";
   const doctorStatus = useDoctorStatusSummary();
   const terminalFallbackPath =
@@ -94,6 +152,124 @@ export function SystemSettings() {
   useEffect(() => {
     void refreshBbCliStatus();
   }, [refreshBbCliStatus]);
+
+  useEffect(() => {
+    let cancelled = false;
+
+    async function loadAppInfo() {
+      if (!window.__TAURI_INTERNALS__) {
+        return;
+      }
+
+      try {
+        const { getIdentifier, getTauriVersion, getVersion } = await import(
+          "@tauri-apps/api/app"
+        );
+        const [version, tauriVersion, identifier] = await Promise.all([
+          getVersion(),
+          getTauriVersion(),
+          getIdentifier(),
+        ]);
+
+        if (!cancelled) {
+          setAppInfo({ version, tauriVersion, identifier });
+        }
+      } catch {
+        if (!cancelled) {
+          setAppInfo(null);
+        }
+      }
+    }
+
+    void loadAppInfo();
+
+    return () => {
+      cancelled = true;
+    };
+  }, []);
+
+  const aboutFallback = t("about.unavailable");
+
+  async function handleLogout() {
+    setLoggingOut(true);
+    try {
+      const nextStatus = await logout();
+      // `onLoggedOut` flips the auth gate, which unmounts AppShell and remounts
+      // it on the next login. Clear the startup latch first so that remount
+      // re-runs startup instead of reusing this account's run.
+      resetChatRuntimeStartup();
+      toast.success(t("account.logoutSuccess"));
+      onLoggedOut?.(nextStatus);
+    } catch (error) {
+      console.warn("Failed to log out:", error);
+      toast.error(t("account.logoutError"));
+    } finally {
+      setLoggingOut(false);
+    }
+  }
+
+  const loadWorkspaces = useCallback(async () => {
+    if (authStatus?.loggedIn !== true) return;
+
+    setWorkspaceLoading(true);
+    setWorkspaceError(false);
+    try {
+      setWorkspaceList(await listAuthWorkspaces());
+    } catch (error) {
+      console.warn("Failed to list workspaces:", error);
+      setWorkspaceError(true);
+    } finally {
+      setWorkspaceLoading(false);
+    }
+  }, [authStatus?.loggedIn]);
+
+  useEffect(() => {
+    void loadWorkspaces();
+  }, [loadWorkspaces]);
+
+  async function handleWorkspaceSwitch(workspaceIdentifier: string) {
+    if (
+      workspaceSwitching ||
+      workspaceIdentifier === workspaceList?.activeWorkspaceIdentifier
+    ) {
+      return;
+    }
+
+    setWorkspaceSwitching(true);
+    try {
+      const result = await switchAuthWorkspace(workspaceIdentifier);
+      const activeWorkspaceIdentifier =
+        result.workspace.workspaceIdentifier ?? workspaceIdentifier;
+      setWorkspaceList((current) =>
+        current
+          ? {
+              ...current,
+              activeWorkspaceIdentifier,
+            }
+          : current,
+      );
+      await queryClient.invalidateQueries();
+      toast.success(
+        t("account.workspace.switchSuccess", {
+          workspace: result.workspace.displayName ?? activeWorkspaceIdentifier,
+        }),
+      );
+    } catch (error) {
+      console.warn("Failed to switch workspace:", error);
+      toast.error(t("account.workspace.switchError"));
+    } finally {
+      setWorkspaceSwitching(false);
+    }
+  }
+
+  const signedInAs =
+    authStatus?.email ?? authStatus?.name ?? authStatus?.user ?? aboutFallback;
+  const organization = authStatus?.org ?? aboutFallback;
+  const showAccountSection = authStatus?.loggedIn === true;
+  const selectableWorkspaces =
+    workspaceList?.workspaces.filter(
+      (workspace) => workspace.workspaceIdentifier,
+    ) ?? [];
 
   async function handleClearMediaCache() {
     setClearingCache(true);
@@ -197,6 +373,11 @@ export function SystemSettings() {
   return (
     <SettingsPage title={t("nav.system")} contentClassName="space-y-8">
       <SettingsSections>
+        {/* Check for updates, first (rev 5): the row people open System for
+          most often. Renders app version + the release channel picker
+          alongside it, same content as the old standalone Updates card. */}
+        {updatesEnabled ? <UpdatesSettings embedded /> : null}
+
         <SettingsSection>
           <SettingsRow
             label={t("general.language.label")}
@@ -413,6 +594,120 @@ export function SystemSettings() {
             ) : null}
 
             {import.meta.env.DEV ? <RuntimeConfigSettings /> : null}
+          </SettingsSection>
+        ) : null}
+
+        {/* About (rev 5): app identity rows + Account, moved here from the
+          standalone About page and pushed to the bottom of System, under
+          their own subhead. App version normally isn't duplicated here
+          since the embedded Updates card above already shows it -- but in
+          updater-disabled builds (VITE_UPDATER_ENABLED=false / capability
+          off) that card doesn't render at all, and this is still the only
+          app-identity surface, so restricted/custom builds would otherwise
+          have no visible version anywhere. Show the row only when the card
+          is absent. Caught by Builderbot review (carried over from About). */}
+        <SettingsSection title={t("about.title")}>
+          {!updatesEnabled ? (
+            <AboutInfoRow
+              label={t("about.fields.version")}
+              value={appInfo?.version ?? aboutFallback}
+            />
+          ) : null}
+          <AboutInfoRow
+            label={t("about.fields.buildMode")}
+            value={
+              import.meta.env.DEV
+                ? t("about.buildModes.development")
+                : t("about.buildModes.production")
+            }
+          />
+          <AboutInfoRow
+            label={t("about.fields.tauriVersion")}
+            value={appInfo?.tauriVersion ?? aboutFallback}
+          />
+          <AboutInfoRow
+            label={t("about.fields.identifier")}
+            value={appInfo?.identifier ?? aboutFallback}
+          />
+          <AboutInfoRow label={t("about.fields.license")} value="Apache-2.0" />
+        </SettingsSection>
+
+        {showAccountSection ? (
+          <SettingsSection title={t("account.title")}>
+            <SettingsRow
+              label={t("account.signedInAs")}
+              description={signedInAs}
+              align="start"
+            >
+              <Button
+                type="button"
+                variant="outline"
+                size="sm"
+                feedbackState={loggingOut ? "loading" : "idle"}
+                loadingLabel={t("account.loggingOut")}
+                disabled={loggingOut}
+                onClick={() => void handleLogout()}
+              >
+                {t("account.logout")}
+              </Button>
+            </SettingsRow>
+            <AboutInfoRow
+              label={t("account.organization")}
+              value={organization}
+            />
+            <SettingsRow
+              label={t("account.workspace.label")}
+              description={
+                workspaceError
+                  ? t("account.workspace.loadError")
+                  : t("account.workspace.description")
+              }
+            >
+              {workspaceError ? (
+                <Button
+                  type="button"
+                  variant="outline"
+                  size="sm"
+                  disabled={workspaceLoading}
+                  onClick={() => void loadWorkspaces()}
+                >
+                  {t("account.workspace.retry")}
+                </Button>
+              ) : (
+                <Select
+                  value={workspaceList?.activeWorkspaceIdentifier ?? undefined}
+                  disabled={
+                    workspaceLoading ||
+                    workspaceSwitching ||
+                    selectableWorkspaces.length === 0
+                  }
+                  onValueChange={(value) => void handleWorkspaceSwitch(value)}
+                >
+                  <SelectTrigger
+                    className="w-52"
+                    aria-label={t("account.workspace.label")}
+                  >
+                    <SelectValue
+                      placeholder={
+                        workspaceLoading
+                          ? t("account.workspace.loading")
+                          : t("account.workspace.unavailable")
+                      }
+                    />
+                  </SelectTrigger>
+                  <SelectContent>
+                    {selectableWorkspaces.map((workspace) => (
+                      <SelectItem
+                        key={workspace.workspaceIdentifier}
+                        value={workspace.workspaceIdentifier ?? ""}
+                      >
+                        {workspace.displayName ?? workspace.workspaceIdentifier}
+                      </SelectItem>
+                    ))}
+                  </SelectContent>
+                </Select>
+              )}
+            </SettingsRow>
           </SettingsSection>
         ) : null}
       </SettingsSections>
