+++ src/shared/i18n/locales/en/settings.json
@@ -405,7 +405,6 @@
     "title": "General"
   },
   "nav": {
-    "about": "About",
     "appearance": "Appearance",
     "archive": "Archive",
     "behavior": "Behavior",
