+++ src/shared/i18n/locales/es/settings.json
@@ -405,7 +405,6 @@
     "title": "General"
   },
   "nav": {
-    "about": "Acerca de",
     "appearance": "Apariencia",
     "archive": "Archivo",
     "behavior": "Comportamiento",
